Extract reindex redaction which Elastic1 applies to non initial index requests and apply equally to Elastic6.
Correctly merge source and update.
