+++ thrall/app/lib/ElasticImageUpdate.scala
@@ -1,0 +1,20 @@
+package lib
+
+import play.api.libs.json.{JsObject, JsValue, Reads, __}
+
+trait ElasticImageUpdate {
+
+  def asImageUpdate(image: JsValue): JsValue = {
+    def removeUploadInformation: Reads[JsObject] =
+      (__ \ "uploadTime").json.prune andThen
+        (__ \ "userMetadata").json.prune andThen
+        (__ \ "exports").json.prune andThen
+        (__ \ "uploadedBy").json.prune andThen
+        (__ \ "collections").json.prune andThen
+        (__ \ "leases").json.prune andThen
+        (__ \ "usages").json.prune
+
+    image.transform(removeUploadInformation).get
+  }
+
+}
